Turned off traverse method for checking remote
The traverse method for checking the existance of files in a remote is
not working properly against ADLSv2. I turned it off by commenting out
the code path that leads to using the traverse method until it can be
fixed.

I added a tertiary version to distiguish it from the existing version.